refactor: change payment class for primary key to be string

+++ database/migrations/2024_05_08_122438_create_payment_methods_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('slug');
-            $table->integer('discount_percentage');
+            $table->float('discount_percentage');
             $table->timestamps();
         });
     }
